requirements.txt
Flask>=2.3.0
Pillow>=10.0.0
pandas>=2.0.0
python-docx>=0.8.11
requests>=2.31.0
